raskatka done!

+++ train_sweep.py
@@ -368,8 +368,8 @@ def train_one_epoch(
             )
     for set_name, batch, _ in bar:
         step += 1
-        inputs = batch['inputs'].to(device)
-        labels = batch['labels'].to(device)
+        inputs = batch['inputs'].to(device).long()
+        labels = batch['labels'].to(device).long()
 
 
         final_logits, logits_steps = model(inputs, return_all_logits=True)
@@ -455,8 +455,8 @@ def evaluate(
                 dynamic_ncols=True
             )
     for set_name, batch, _ in bar:
-        inputs = batch['inputs'].to(device)
-        labels = batch['labels'].to(device)
+        inputs = batch['inputs'].to(device).long()
+        labels = batch['labels'].to(device).long()
 
         final_logits, logits_steps = model(inputs, return_all_logits=True)
